requirements.txt
pyocr
pillow
spacy
git+https://github.com/aparrish/pytracery.git#egg=tracery
